Move IPC related SVCs into a separate file, and logging on RegisterService (sm)
